Move `RWCounter` to bus-mapping operation.
